requirements.txt
google-genai
python-dotenv
jiwer
evaluate
rouge-score